requirements.txt
alembic==1.15.2
anyio==4.9.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
google-api-core==2.25.1
google-api-python-client==2.116.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
googlesearch-python==1.2.3
greenlet==3.2.1
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
jieba3k==0.35.1
joblib==1.5.1
lxml==5.2.1
lxml_html_clean==0.1.1
Mako==1.3.10
MarkupSafe==3.0.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# Computer Vision
opencv-python
torch
transformers
pillow
openai
supabase
python-dotenv